requirements_inference.txt
av
paddlepaddle
tqdm
pims